requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
Django==4.0.4
django-bootstrap-v5==1.0.11
django-stdimage==6.0.1
gunicorn==20.1.0
Pillow==9.1.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
whitenoise==6.1.0
